fix: created fish component
