+++ lib/aquarium.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:gussuri/component/Fish.dart';
 
 class Aquarium extends StatefulWidget {
   const Aquarium({Key? key}) : super(key: key);
@@ -7,60 +8,21 @@ class Aquarium extends StatefulWidget {
   _AquariumState createState() => _AquariumState();
 }
 
-class _AquariumState extends State<Aquarium>
-    with SingleTickerProviderStateMixin {
-  late AnimationController controller;
-
-  @override
-  void initState() {
-    controller =
-        AnimationController(duration: const Duration(seconds: 2), vsync: this)
-          ..forward()
-          ..addListener(() {
-            if (controller.isCompleted) {
-              controller.repeat();
-            }
-          });
-    super.initState();
-  }
-
-  @override
-  void dispose() {
-    controller.dispose();
-    super.dispose();
-  }
-
-  double shake(double value) =>
-      1 * (0.5 - (0.5 - Curves.bounceOut.transform(value)).abs());
-
+class _AquariumState extends State<Aquarium> {
 
   @override
   Widget build(BuildContext context) {
 
-    return Container(
-      child: Stack(
-        children: [
-          Container(
-              decoration: const BoxDecoration(
-                color: Colors.black,
-                  image: DecorationImage(
-                      image: AssetImage('images/background.jpg'), fit: BoxFit.cover)),
-          ),
-          SafeArea(child:  AnimatedBuilder(
-            animation: controller,
-            builder: (context, child) => Transform.translate(
-              offset: Offset(200 * shake(controller.value), 0),
-              child: child,
-            ),
-            child: const Image(
-              image: AssetImage('images/fish.gif'),
-              height: 100,
-              width: 100,
-              fit: BoxFit.cover,
-            ),
-          ))
-        ],
-      ),
+    return Stack(
+      children: [
+        Container(
+            decoration: const BoxDecoration(
+              color: Colors.black,
+                image: DecorationImage(
+                    image: AssetImage('images/background.jpg'), fit: BoxFit.cover)),
+        ),
+        Fish(),
+      ],
     );
   }
 }
